added faster tile spawner+adjacent mine minimizer
Using an array of all possible positions on the minefield I can randomly select an index holding a position value and push it onto an array of mines. The I can set the position on the initial array to the end index and pop that array. Using this method to randomly select positions, there is no possible bottleneck of trying to select a random position that already has a mine on it.